+++ sea/server/__init__.py
@@ -1,0 +1,1 @@
+from .threading import Server as Server  # noqa