+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
 ## [Unreleased]
+
+### Added
+- Added multiprocessing worker class
 ### Changed
 - Use cached_property from standard lib when possible
+
 
 ## [2.3.3] - 2022-04-22
 ### Added